cgen: dump: refactor the program dump logic
The program dump logic is buggy and incomplete, leading to erroneous
translation of the program. Maintaining such a logic is cumbersome as
we need to handle various edge cases.

Inspired by bpftool, the program dump logic is replaced by Linux
disasm.c from the kernel tree. Thanks change ensures the printed
bytecode is consistent with the output of the BPF verifier and bpftool.

Thanks to Quentin Monnet his advice to perform the bytecode dump.