daemon: ctx: identify codegens by their chain's name
Chains are expected to have unique (user-defined) names. Modify the
context to store all the codegens into a single list and retrieve them
based on their chain's unique name.

Thanks to the chain's name, there is no need to have an array of bf_list
to store codegens based on their hook, all the codegens can be stored in
a single bf_list.