cgen: split load and attachment steps
Instead of loading and attaching a BPF program in a single step, split
those in two: loading and attachment. This change will allow users to
define a new chain without attaching it, instead of relying on attach=no
option.

Additionally, only define a single bf_link in bf_program. This is
sufficient for now, even though it will force users to define 2 chains
to filter on Netfilter hooks for IPv4 and IPv6. However, it makes it
simpler to handle errors and update programs atomically.