lib: add API to retrieve a chain program's FD
bf_chain_prog_fd() will return the file descriptor of a program based on
the chain's name. If no program exists, an error is returned. The caller
is reponsible for closing the file descriptor. This doesn't prevent
bpfilter from re-generating the program, although it doesn't invalidate
the client's file descriptor.