daemon: cgen: introduce bf_handle to store BPF objects references
bf_program is used to generate the bytecode and handle reference to the
generated BPF objects (program, link, maps). bf_program shouldn't be a
long term object, as we don't need to keep the bytecode around.

bf_handle is a structure aimed to store reference to the BPF objects,
and restore the references from a given BPF object ID. It will allow
bf_program to focus on bytecode generation, and remove the need to
serialize bf_program.